+++ src/scripts/pcap_log-converter.sh
@@ -6,5 +6,8 @@ if ! command -v tshark; then
   exit 1
 fi
 
+# We want tshark output to come in UTC
+export TZ=UTC
+
 # Use tshark to convert the pcap file into a JSON-lines log file
 exec tshark -T ek -P -V -t ad -r $2